% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltasurvey-package.R
\docType{package}
\name{deltasurvey-package}
\alias{deltasurvey}
\alias{deltasurvey-package}
\title{deltasurvey: Functions for the California Delta Residents Survey Data}
\description{
Facilitates the proper loading of California Delta Residents Survey data (2023) available on Open ICPSR, with the DOI provided in 'DataSet'. This includes reading the data into the R environment correctly.
}
\author{
\strong{Maintainer}: Kenji Tomari \email{ktomari@ucdavis.edu} (\href{https://orcid.org/0000-0001-8039-3999}{ORCID})

}
\keyword{internal}
