% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{drs_as_NA}
\alias{drs_as_NA}
\title{Convert columns with missing values in the form of <angle brackets> to \code{NA}.}
\usage{
drs_as_NA(.data)
}
\arguments{
\item{.data}{A data.frame or tibble consisting of the data from the DRS.}
}
\value{
A data.frame or tibble.
}
\description{
Convert columns with missing values in the form of <angle brackets> to \code{NA}.
}
\examples{
df <- data.frame(Q0_0 = c("1", "<Missing>", "3"))
drs_as_NA(df)
}
